+++ Changes.md
@@ -5,6 +5,7 @@ Fixes
 -----
 
 - 3Delight : Fixed rendering with recent versions of 3Delight, which no longer use the `DL_DISPLAYS_PATH` to search for display drivers.
+- Metadata : Fixed memory leak.
 
 API
 ---